requirements.txt
fastapi
google-auth 
google-auth-oauthlib 
google-api-python-client
uvicorn[standard]
itsdangerous
starlette
msal